remove cupy/cuda/_environment.py
